refactor: use ability utilities
